requirements.txt
fastapi
uvicorn
requests
pydantic
motor
python-dotenv